Buffers

Buffers

ReinforcementLearning.ArrayCircularBuffer — Type.

mutable struct ArrayCircularBuffer{T}
    data::T
    capacity::Int64
    start::Int64
    counter::Int64
    full::Bool

source

ReinforcementLearning.ArrayCircularBuffer — Method.

ArrayCircularBuffer(arraytype, datatype, elemshape, capacity)

source

ReinforcementLearning.ArrayStateBuffer — Type.

struct ArrayStateBuffer{Ts, Ta}
    states::ArrayCircularBuffer{Ts}
    actions::CircularBuffer{Ta}
    rewards::CircularBuffer{Float64}
    done::CircularBuffer{Bool}

source

ReinforcementLearning.ArrayStateBuffer — Method.

ArrayStateBuffer(; arraytype = Array, datatype = Float64, 
                   elemshape = (1), actiontype = Int64, 
                   capacity = 2, capacitystates = capacity,
                   capacityrewards = capacity - 1)

An ArrayStateBuffer is similar to a Buffer but the states are stored in a prealocated array of size (elemshape..., capacity). K consecutive states at position i in the state buffer can can efficiently be retrieved with nmarkovview(buffer.states, i, K) or nmarkovgetindex(buffer.states, i, K). See the implementation of DQN for an example.

source

ReinforcementLearning.Buffer — Type.

struct Buffer{Ts, Ta}
    states::CircularBuffer{Ts}
    actions::CircularBuffer{Ta}
    rewards::CircularBuffer{Float64}
    done::CircularBuffer{Bool}

source

ReinforcementLearning.Buffer — Method.

Buffer(; statetype = Int64, actiontype = Int64, 
         capacity = 2, capacitystates = capacity,
         capacityrewards = capacity - 1)

source

ReinforcementLearning.EpisodeBuffer — Type.

struct EpisodeBuffer{Ts, Ta}
    states::Array{Ts, 1}
    actions::Array{Ta, 1}
    rewards::Array{Float64, 1}
    done::Array{Bool, 1}

source

ReinforcementLearning.EpisodeBuffer — Method.

EpisodeBuffer(; statetype = Int64, actiontype = Int64) = 
    EpisodeBuffer(statetype[], actiontype[], Float64[], Bool[])

source