Callbacks

Callbacks

ReinforcementLearning.AllRewards — Type.

struct AllRewards
    rewards::Array{Float64, 1}

Records all rewards.

source

ReinforcementLearning.AllRewards — Method.

AllRewards()

Initializes with empty array.

source

ReinforcementLearning.EvaluateGreedy — Method.

EvaluateGreedy(callback, stoppincriterion; every = Episode(10))

Evaluate an rlsetup greedily by leaving the normal learning loop and evaluating the agent with callback until stoppingcriterion is met, at which point normal learning is resumed. This is done every Nth Episode (where N = 10 by default) or every Nth Step (e.g. every = Step(10)).

Example:

eg = EvaluateGreedy(EvaluationPerEpisode(TotalReward(), returnmean = true),
                    ConstantNumberEpisodes(10), every = Episode(100))
rlsetup = RLSetup(learner, environment, stoppingcriterion, callbacks = [eg])
learn!(rlsetup)
getvalue(eg)

Leaves the learning loop every 100th episode to estimate the average total reward per episode, by running a greedy policy for 10 episodes.

source

ReinforcementLearning.LinearDecreaseEpsilon — Type.

mutable struct LinearDecreaseEpsilon
    start::Int64
    stop::Int64
    initval::Float64
    finalval::Float64
    t::Int64
    step::Float64

Linearly decrease ϵ of an EpsilonGreedyPolicy from initval until step start to finalval at step stop.

Stepsize step = (finalval - initval)/(stop - start).

source

ReinforcementLearning.LinearDecreaseEpsilon — Method.

LinearDecreaseEpsilon(start, stop, initval, finalval)

source

ReinforcementLearning.Progress — Type.

mutable struct Progress 
    steps::Int64
    laststopcountervalue::Int64

Show steps times progress information during learning.

source

ReinforcementLearning.Progress — Type.

Progress(steps = 10) = Progress(steps, 0)

source

ReinforcementLearning.RecordAll — Type.

struct RecordAll
    rewards::Array{Float64, 1}
    actions::Array{Int64, 1}
    states::Array{Int64, 1}
    done::Array{Bool, 1}

Records everything.

source

ReinforcementLearning.RecordAll — Method.

RecordAll()

Initializes with empty arrays.

source

ReinforcementLearning.ReduceEpsilonPerEpisode — Type.

mutable struct ReduceEpsilonPerEpisode
    ϵ0::Float64
    counter::Int64

Reduces ϵ of an EpsilonGreedyPolicy after each episode.

In episode n, ϵ = ϵ0/n

source

ReinforcementLearning.ReduceEpsilonPerEpisode — Method.

ReduceEpsilonPerEpisode()

Initialize callback.

source

ReinforcementLearning.ReduceEpsilonPerT — Type.

mutable struct ReduceEpsilonPerT
    ϵ0::Float64
    T::Int64
    n::Int64
    counter::Int64

Reduces ϵ of an EpsilonGreedyPolicy after every T steps.

After n * T steps, ϵ = ϵ0/n

source

ReinforcementLearning.ReduceEpsilonPerT — Method.

ReduceEpsilonPerT()

Initialize callback.

source

ReinforcementLearning.Visualize — Type.

mutable struct Visualize 
    plot
    wait::Float64

source

ReinforcementLearning.Visualize — Method.

Visualize(; wait = .15)

A callback to be used in an RLSetup to visualize an environment during running or learning.

source