Comparison

Comparison Tools

ReinforcementLearning.compare — Method.

compare(rlsetupcreators::Dict, N; callbackid = 1, verbose = false)

Run different setups in dictionary rlsetupcreators N times. The dictionary has elements "name" => createrlsetup, where createrlsetup is a function that has a single integer argument (id of the comparison; useful for saving intermediate results). For each run, getvalue(rlsetup.callbacks[callbackid]) gets entered as result in a DataFrame with columns "name", "result", "seed". It is useful to specify the callbackid if the createrlsetup creates RLSetups with multiple callbacks.

Example

using ReinforcementLearningEnvironmentDiscrete
env = MDP()
setup(learner) = RLSetup(learner, env, ConstantNumberSteps(10^4), callbacks = [EvaluationPerT(10^2, MeanReward())])
rlsetupcreators = Dict("sarsa" => (i) -> setup(Sarsa()), "smallbackups" => (i) -> setup(SmallBackups()))
result = compare(rlsetupcreators, 4)
plotcomparison(result)

source

ReinforcementLearning.plotcomparison — Method.

plotcomparison(df; nmaxpergroup = 20, linestyles = [], 
                   showbest = true, axisoptions = @pgf {})

Plots results obtained with compare using PGFPlotsX.

source