Action Selectors

Action Selectors

ReinforcementLearning.AbstractDiscreteActionSelector — Type.

AbstractDiscreteActionSelector

A subtype of AbstractDiscreteActionSelector is used to generate an action given the estimated value of different actions.

Required Methods	Brief Description

selector(values;step, kw...)	selector, an instance of AbstractDiscreteActionSelector, must be a callable object which takes in an estimation and returns an action.

source

ReinforcementLearning.AlternateSelector — Type.

AlternateSelector <: AbstractDiscreteActionSelector

Used to ensure that all actions are selected alternatively.

AlternateSelector(n::Int)

n::Int means the optional actions are 1:n.

source

ReinforcementLearning.EpsilonGreedySelector — Type.

(s::EpsilonGreedySelector)(values; step) where T

Note

If multiple values with the same maximum value are found. Then a random one will be returned!

NaN will be filtered unless all the values are NaN. In that case, a random one will be returned.

source

ReinforcementLearning.UCBSelector — Type.

UCBSelector <: AbstractDiscreteActionSelector
UCBSelector(na; c=2.0, ϵ=1e-10)

Arguments

na is the number of actions used to create a internal counter.

t is used to store current time step.

c is used to control the degree of exploration.

source

ReinforcementLearning.WeightedSelector — Type.

WeightedSelector <: AbstractDiscreteActionSelector
WeightedSelector(is_normalized::Bool)

is_normalized is used to indicating if the feeded action values are alrady normalized to have a sum of 1.0.

source