Environment Models

ReinforcementLearning.AbstractEnvironmentModel — Type

Describe how to model a reinforcement learning environment.

See also AbstractDistributionBasedModel, AbstractSampleBasedModel.

ReinforcementLearning.AbstractSampleBasedModel — Type

A collection of models that can be used to sample transitions.

ReinforcementLearning.AbstractDistributionBasedModel — Type

A collection of models that can get the distribution given a state and an actioi.

ReinforcementLearning.DynamicDistributionModel — Type

DynamicDistributionModel(f::Tf, ns::Int, na::Int) -> DynamicDistributionModel{Tf}

Use a general function f to store the transformations. ns and na are the number of states and actions.

ReinforcementLearning.DeterministicDistributionModel — Type

DeterministicDistributionModel(table::Array{Vector{NamedTuple{(:nextstate, :reward, :prob),Tuple{Int,Float64,Float64}}}, 2})

Store all the transformations in the table field.

ReinforcementLearning.ExperienceBasedSampleModel — Type

ExperienceBasedSampleModel() -> ExperienceBasedSampleModel

Generate a transition based on previous experiences.

ReinforcementLearning.PrioritizedSweepingSampleModel — Type

PrioritizedSweepingSampleModel(θ::Float64=1e-4)

See more details at Section (8.4) on Page 168 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

ReinforcementLearning.TimeBasedSampleModel — Type

TimeBasedSampleModel(nactions::Int, κ::Float64 = 1e-4)