Environments

Environments

This package relies on some interfaces provided by the ReinforcementLearningEnvironments.jl (RLEnvs). For completeness, we will also give a short introduction to it here.

RLEnvs provides many interfaces similar to OpenAI Gym. But also extends it a little bit to make things easier to interact with sync/async, single/multi agent environments.

Basically, an environment is a functional object which takes in an action and changes its internal state correspondly. For sync environments, both env(action) and reset!(env) should return nothing, and observe(env) should return an Observation. For async environments, they should all return a task.

A specially kind of environment is WrappedEnv.

Missing docstring.

Missing docstring for Observation. Check Documenter's build log for details.

ReinforcementLearning.WrappedEnv — Type.

WrappedEnv(;env, preprocessor)

The observation of env is first processed by the preprocessor.

source

Preprocessors

Following are some built-in preprocessors. Notice that preprocessors can be chained (like Chain(p1, p2, ps...)) to get a composed preprocessor.

ReinforcementLearning.AbstractPreprocessor — Type.

Preprocess an Observation and return a new observation. By default, the preprocessor is only applied to the state field of the Observation and other fields remain unchanged.

For customized preprocessors that inherit AbstractPreprocessor, you can change this behavior by rewriting (p::AbstractPreprocessor)(obs::Observation) method.

source

ReinforcementLearning.FourierPreprocessor — Type.

FourierPreprocessor(order::Int)

Transform a scalar to a vector of order+1 Fourier bases.

source

ReinforcementLearning.PolynomialPreprocessor — Type.

PolynomialPreprocessor(order::Int)

Transform a scalar to vector of maximum order polynomial.

source

ReinforcementLearning.TilingPreprocessor — Type.

TilingPreprocessor(tilings::Vector{<:Tiling})

Use each tilings to encode the state and return a vector.

source

ReinforcementLearning.ImageResize — Type.

ImageResize(img::Array{T, N})
ImageResize(dims::Int...) -> ImageResize(Float32, dims...)
ImageResize(T::Type{<:Number}, dims::Int...)

Using BSpline method to resize the state field of an Observation to size of img (or dims).

source

ReinforcementLearning.ImageCrop — Type.

struct ImageCrop
    xidx::UnitRange{Int64}
    yidx::UnitRange{Int64}
end

Select indices xidx and yidx from a 2 or 3 dimensional array. TODO: Inefficient!

Example:

c = ImageCrop(2:5, 3:2:9)
c([10i + j for i in 1:10, j in 1:10])

source

ReinforcementLearning.ImageResizeNearestNeighbour — Type.

struct ImageResizeNearestNeighbour
    outdim::Tuple{Int64, Int64}
end

Resize any image to outdim = (width, height) by nearest-neighbour interpolation (i.e. subsampling).

Example:

r = ImageResizeNearestNeighbour((50, 50))
r(rand(200, 200))
r(rand(UInt8, 3, 100, 100))

source