Core

Core

Base.run — Function.

run(agent::AbstractAgent, env::AbstractEnv, stop_condition; hook = EmptyHook())

See also Stop Conditions, Hooks

source

ReinforcementLearning.policy_evaluation! — Function.

policy_evaluation!(V::AbstractVApproximator, π, model::AbstractDistributionBasedModel; γ::Float64=0.9, θ::Float64=1e-4)

See more details at Section (4.1) on Page 75 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

ReinforcementLearning.policy_improvement! — Function.

policy_improvement!(;V::AbstractVApproximator, π::AbstractPolicy, model::AbstractDistributionBasedModel, γ::Float64 = 0.9)

See more details at Section (4.2) on Page 76 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

ReinforcementLearning.policy_iteration! — Function.

policy_iteration!(V::AbstractVApproximator, π, model::AbstractDistributionBasedModel; γ::Float64=0.9, θ::Float64=1e-4, max_iter=typemax(Int))

See more details at Section (4.3) on Page 80 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

ReinforcementLearning.value_iteration! — Function.

value_iteration!(V::AbstractVApproximator, model::AbstractDistributionBasedModel; γ::Float64=0.9, θ::Float64=1e-4, max_iter=typemax(Int))

See more details at Section (4.4) on Page 83 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Stop Conditions

ReinforcementLearning.StopAfterStep — Type.

StopAfterStep(step; cur = 1, is_show_progress = true, tag = "TRAINING")

Return true after being called for step.

source

ReinforcementLearning.StopAfterEpisode — Type.

StopAfterEpisode(episode; cur = 0, is_show_progress = true, tag = "TRAINING")

Return true after being called episode. If is_show_progress is true, the ProgressMeter will be used to show progress.

source

ReinforcementLearning.StopWhenDone — Type.

StopWhenDone()

Return true if the terminal field of an observation is true.

source

ReinforcementLearning.ComposedStopCondition — Type.

ComposedStopCondition(stop_conditions; reducer = any)

The result of stop_conditions is reduced by reducer.

source

Hooks

ReinforcementLearning.AbstractHook — Type.

A hook is called at different stage duiring a run. One can inject customized runtime logic in it.

source

ReinforcementLearning.ComposedHook — Type.

ComposedHook(hooks...)

Compose different hooks into a single hook.

source

ReinforcementLearning.EmptyHook — Type.

Do nothing

source

ReinforcementLearning.StepsPerEpisode — Type.

StepsPerEpisode(; steps = Int[], count = 0, tag = "TRAINING")

Store steps of each episode in the field of steps.

source

ReinforcementLearning.RewardsPerEpisode — Type.

RewardsPerEpisode(; rewards = Vector{Vector{Float64}}(), tag = "TRAINING")

Store each reward of each step in every episode in the field of rewards.

source

ReinforcementLearning.TotalRewardPerEpisode — Type.

TotalRewardPerEpisode(; rewards = Float64[], reward = 0.0, tag = "TRAINING")

Store the total rewards of each episode in the field of rewards.

source

ReinforcementLearning.CumulativeReward — Type.

CumulativeReward(rewards::Vector{Float64} = [0.0], tag::String = "TRAINING")

Store cumulative rewards since the beginning to the field of rewards.

source

ReinforcementLearning.TimePerStep — Type.

TimePerStep(;max_steps=100)
TimePerStep(times::CircularArrayBuffer{Float64, 1}, t::UInt64)

Store time cost of the latest max_steps in the times field.

source