Learners

Learners

TD Learner

ReinforcementLearning.ExpectedSarsa — Method.

ExpectedSarsa(; kargs...) = TDLearner(; endvaluepolicy = ExpectedSarsaEndPolicy(VeryOptimisticEpsilonGreedyPolicy(.1)), kargs...)

source

ReinforcementLearning.QLearning — Method.

QLearning(; kargs...) = TDLearner(; endvaluepolicy = QLearningEndPolicy(), kargs...)

source

ReinforcementLearning.Sarsa — Method.

Sarsa(; kargs...) = TDLearner(; kargs...)

source

ReinforcementLearning.AccumulatingTraces — Type.

struct AccumulatingTraces <: AbstractTraces
    λ::Float64
    γλ::Float64
    trace::Array{Float64, 2}
    minimaltracevalue::Float64

Decaying traces with factor γλ.

Traces are updated according to $e(a, s) ← 1 + e(a, s)$ for the current action-state pair and $e(a, s) ← γλ e(a, s)$ for all other pairs unless $e(a, s) <$ minimaltracevalue where the trace is set to 0 (for computational efficiency).

source

ReinforcementLearning.AccumulatingTraces — Method.

AccumulatingTraces(ns, na, λ::Float64, γ::Float64; minimaltracevalue = 1e-12)

source

ReinforcementLearning.NoTraces — Type.

struct NoTraces <: AbstractTraces

No eligibility traces, i.e. $e(a, s) = 1$ for current action $a$ and state $s$ and zero otherwise.

source

ReinforcementLearning.ReplacingTraces — Type.

struct ReplacingTraces <: AbstractTraces
    λ::Float64
    γλ::Float64
    trace::Array{Float64, 2}
    minimaltracevalue::Float64

Decaying traces with factor γλ.

Traces are updated according to $e(a, s) ← 1$ for the current action-state pair and $e(a, s) ← γλ e(a, s)$ for all other pairs unless $e(a, s) <$ minimaltracevalue where the trace is set to 0 (for computational efficiency).

source

ReinforcementLearning.ReplacingTraces — Method.

ReplacingTraces(ns, na, λ::Float64, γ::Float64; minimaltracevalue = 1e-12)

source

ReinforcementLearning.TDLearner — Type.

mutable struct TDLearner{T,Tp}
    ns::Int64 = 10
    na::Int64 = 4
    γ::Float64 = .9
    λ::Float64 = .8
    α::Float64 = .1
    nsteps::Int64 = 1
    initvalue::Float64 = 0.
    unseenvalue::Float64 = initvalue == Inf64 ? 0. : initvalue
    params::Array{Float64, 2} = zeros(na, ns) .+ initvalue
    tracekind = DataType = λ == 0 ? NoTraces : ReplacingTraces
    traces::T = tracekind == NoTraces ? NoTraces() : tracekind(ns, na, λ, γ)
    endvaluepolicy::Tp = SarsaEndPolicy()

source

Initial values, novel actions and unseen values

For td-error dependent methods, the exploration-exploitation trade-off depends on the initvalue and the unseenvalue. To distinguish actions that were never choosen before, i.e. novel actions, the default initial Q-value (field param) is initvalue = Inf64. In a state with novel actions, the policy determines how to deal with novel actions. To compute the td-error the unseenvalue is used for states with novel actions. One way to achieve agressively exploratory behavior is to assure that unseenvalue (or initvalue) is larger than the largest possible Q-value.

Policy Gradient Learner

ReinforcementLearning.Critic — Type.

mutable struct Critic <: AbstractBiasCorrector
    α::Float64
    V::Array{Float64, 1}

source

ReinforcementLearning.Critic — Method.

Critic(; γ = .9, α = .1, ns = 10, initvalue = 0.)

source

ReinforcementLearning.NoBiasCorrector — Type.

struct NoBiasCorrector <: AbstractBiasCorrector

source

ReinforcementLearning.RewardLowpassFilterBiasCorrector — Type.

mutable struct RewardLowpassFilterBiasCorrector <: AbstractBiasCorrector
λ::Float64
rmean::Float64

Filters the reward with factor λ and uses effective reward (r - rmean) to update the parameters.

source

ReinforcementLearning.ActorCriticPolicyGradient — Method.

ActorCriticPolicyGradient(; nsteps = 1, γ = .9, ns = 10, na = 4, 
                            α = .1, αcritic = .1, initvalue = Inf64)

source

ReinforcementLearning.EpisodicReinforce — Method.

EpisodicReinforce(; kwargs...) = PolicyGradientForward(; kwargs...)

source

ReinforcementLearning.AbstractPolicyGradient — Type.

mutable struct PolicyGradientBackward <: AbstractPolicyGradient
    ns::Int64 = 10
    na::Int64 = 4
    γ::Float64 = .9
    α::Float64 = .1
    initvalue::Float64 = 0.
    params::Array{Float64, 2} = zeros(na, ns) + initvalue
    traces::AccumulatingTraces = AccumulatingTraces(ns, na, 1., γ, 
                                                    trace = zeros(na, ns))
    biascorrector::T = NoBiasCorrector()

Policy gradient learning in the backward view.

The parameters are updated according to $params[a, s] += α * r_{eff} * e[a, s]$ where $r_{eff} = r$ for NoBiasCorrector, $r_{eff} = r - rmean$ for RewardLowpassFilterBiasCorrector and e[a, s] is the eligibility trace.

source

ReinforcementLearning.PolicyGradientForward — Type.

mutable struct PolicyGradientForward <: AbstractPolicyGradient
    ns::Int64 = 10
    na::Int64 = 4
    γ::Float64 = .9
    α::Float64 = .1
    initvalue::Float64 = 0.
    params::Array{Float64, 2} = zeros(na, ns) + initvalue
    biascorrector::Tb = NoBiasCorrector()
    nsteps::Int64 = typemax(Int64)

source

N-step Learner

ReinforcementLearning.MonteCarlo — Type.

mutable struct MonteCarlo <: AbstractReinforcementLearner
    ns::Int64 = 10
    na::Int64 = 4
    γ::Float64 = .9
    initvalue = 0.
    Nsa::Array{Int64, 2} = zeros(Int64, na, ns)
    Q::Array{Float64, 2} = zeros(na, ns) + initvalue

Estimate Q values by averaging over returns.

source

Model Based Learner

ReinforcementLearning.MDPLearner — Type.

@with_kw struct MDPLearner
    mdp::MDP = MDP()
    γ::Float64 = .9
    policy::Array{Int64, 1} = ones(Int64, mdp.ns)
    values::Array{Float64, 1} = zeros(mdp.ns)

Used to solve mdp with discount factor γ.

source

ReinforcementLearning.policy_iteration! — Method.

policy_iteration!(mdplearner::MDPLearner)

Solve MDP with policy iteration using MDPLearner.

source

ReinforcementLearning.SmallBackups — Type.

mutable struct SmallBackups <: AbstractReinforcementLearner
    ns::Int64 = 10
    na::Int64 = 4
    γ::Float64 = .9
    initvalue::Float64 = Inf64
    maxcount::UInt64 = 3
    minpriority::Float64 = 1e-8
    M::Int64 = 1
    counter::Int64 = 0
    Q::Array{Float64, 2} = zeros(na, ns) .+ initvalue
    V::Array{Float64, 1} = zeros(ns) .+ (initvalue == Inf64 ? 0. : initvalue)
    U::Array{Float64, 1} = zeros(ns) .+ (initvalue == Inf64 ? 0. : initvalue)
    Nsa::Array{Int64, 2} = zeros(Int64, na, ns)
    Ns1a0s0::Array{Dict{Tuple{Int64, Int64}, Int64}, 1} = [Dict{Tuple{Int64, Int64}, Int64}() for _ in 1:ns]
    queue::PriorityQueue = PriorityQueue(Base.Order.Reverse, zip(Int64[], Float64[]))

See Harm Van Seijen, Rich Sutton ; Proceedings of the 30th International Conference on Machine Learning, PMLR 28(3):361-369, 2013.

maxcount defines the maximal number of backups per action, minpriority is the smallest priority still added to the queue.

source

Deep Reinforcement Learning

ReinforcementLearning.DQN — Type.

mutable struct DQN{Tnet,TnetT,ToptT,Topt}
    γ::Float64 = .99
    na::Int64
    net::TnetT
    targetnet::Tnet = Flux.mapleaves(Flux.Tracker.data, deepcopy(net))
    policynet::Tnet = Flux.mapleaves(Flux.Tracker.data, net)
    updatetargetevery::Int64 = 500
    t::Int64 = 0
    updateevery::Int64 = 1
    opttype::ToptT = Flux.ADAM
    opt::Topt = opttype(Flux.params(net))
    startlearningat::Int64 = 10^3
    minibatchsize::Int64 = 32
    doubledqn::Bool = true
    nmarkov::Int64 = 1
    nsteps::Int64 = 1
    replaysize::Int64 = 10^4
    loss::Function = Flux.mse

source

ReinforcementLearning.DeepActorCritic — Type.

mutable struct DeepActorCritic{Tnet, Tpl, Tplm, Tvl, ToptT, Topt}
    nh::Int64 = 4
    na::Int64 = 2
    γ::Float64 = .9
    nsteps::Int64 = 5
    net::Tnet
    policylayer::Tpl = Linear(nh, na)
    policynet::Tplm = Flux.Chain(Flux.mapleaves(Flux.Tracker.data, net),
                             Flux.mapleaves(Flux.Tracker.data, policylayer))
    valuelayer::Tvl = Linear(nh, 1)
    params::Array{Any, 1} = vcat(map(Flux.params, [net, policylayer, valuelayer])...)
    t::Int64 = 0
    updateevery::Int64 = 1
    opttype::ToptT = Flux.ADAM
    opt::Topt = opttype(params)
    αcritic::Float64 = .1
    nmarkov::Int64 = 1

source