Evaluation Metrics

Evaluation Metrics

ReinforcementLearning.EvaluationPerEpisode — Type.

EvaluationPerEpisode
    values::Array{Float64, 1}
    metric::SimpleEvaluationMetric

Stores the value of the simple metric for each episode in values.

source

ReinforcementLearning.EvaluationPerEpisode — Type.

EvaluationPerEpisode(metric = MeanReward())

Initializes with empty values array and simple metric (default MeanReward). Other options are TimeSteps (to measure the lengths of episodes) or TotalReward.

source

ReinforcementLearning.EvaluationPerT — Type.

EvaluationPerT
    T::Int64
    counter::Int64
    values::Array{Float64, 1}
    metric::SimpleEvaluationMetric

Stores the value of the simple metric after every T steps in values.

source

ReinforcementLearning.EvaluationPerT — Type.

EvaluationPerT(T, metric = MeanReward())

Initializes with T, counter = 0, empty values array and simple metric (default MeanReward). Another option is TotalReward.

source

ReinforcementLearning.MeanReward — Type.

mutable struct MeanReward 
    meanreward::Float64
    counter::Int64

Computes iteratively the mean reward.

source

ReinforcementLearning.MeanReward — Method.

MeanReward()

Initializes counter and meanreward to 0.

source

ReinforcementLearning.TimeSteps — Type.

mutable struct TimeSteps
    counter::Int64

Counts the number of timesteps the simulation is running.

source

ReinforcementLearning.TimeSteps — Method.

TimeSteps()

Initializes counter to 0.

source

ReinforcementLearning.TotalReward — Type.

mutable struct TotalReward 
    reward::Float64

Accumulates all rewards.

source

ReinforcementLearning.TotalReward — Method.

TotalReward()

Initializes reward to 0.

source