Policies

Policies

Epsilon Greedy Policies

ReinforcementLearning.EpsilonGreedyPolicy — Type.

mutable struct EpsilonGreedyPolicy{kind}
    ϵ::Float64

Chooses the action with the highest value with probability 1 - ϵ and selects an action uniformly random with probability ϵ.

source

Softmax Policies

ReinforcementLearning.AbstractSoftmaxPolicy — Type.

mutable struct SoftmaxPolicy <: AbstractSoftmaxPolicy
    β::Float64

Choose action $a$ with probability

\[\frac{e^{\beta x_a}}{\sum_{a'} e^{\beta x_{a'}}}\]

where $x$ is a vector of values for each action. In states with actions that were never chosen before, a uniform random novel action is returned.

SoftmaxPolicy(; β = 1.)

Returns a SoftmaxPolicy with default β = 1.

source

Forced Policy and Episode

ReinforcementLearning.ForcedEpisode — Type.

mutable struct ForcedEpisode{Ts}
    t::Int64
    states::Ts
    dones::Array{Bool, 1}
    rewards::Array{Float64, 1}

source

ReinforcementLearning.ForcedPolicy — Type.

mutable struct ForcedPolicy 
    t::Int64
    actions::Array{Int64, 1}

source