ReinforcementLearningEnvironments.jl

ReinforcementLearningEnvironments.AcrobotEnv — Method

AcrobotEnv(;kwargs...)

Keyword arguments

T = Float64

link_length_a = T(1.)

link_length_b = T(1.)

link_mass_a = T(1.)

link_mass_b = T(1.)

link_com_pos_a = T(0.5)

link_com_pos_b = T(0.5)

link_moi = T(1.)

max_vel_a = T(4 * π)

max_vel_b = T(9 * π)

g = T(9.8)

dt = T(0.2)

max_steps = 200

book_or_nips = 'book'

avail_torque = [T(-1.), T(0.), T(1.)]

ReinforcementLearningEnvironments.CartPoleEnv — Method

CartPoleEnv(;kwargs...)

Keyword arguments

T = Float64

gravity = T(9.8)

masscart = T(1.0)

masspole = T(0.1)

halflength = T(0.5)

forcemag = T(10.0)

max_steps = 200

'dt = 0.02'

rng = Random.GLOBAL_RNG

ReinforcementLearningEnvironments.MountainCarEnv — Method

MountainCarEnv(;kwargs...)

Keyword arguments

T = Float64

continuous = false

rng = Random.GLOBAL_RNG

min_pos = -1.2

max_pos = 0.6

max_speed = 0.07

goal_pos = 0.5

max_steps = 200

goal_velocity = 0.0

power = 0.001

gravity = 0.0025

ReinforcementLearningEnvironments.PendulumEnv — Method

PendulumEnv(;kwargs...)

Keyword arguments

T = Float64

max_speed = T(8)

max_torque = T(2)

g = T(10)

m = T(1)

l = T(1)

dt = T(0.05)

max_steps = 200

continuous::Bool = true

n_actions::Int = 3

rng = Random.GLOBAL_RNG

ReinforcementLearningEnvironments.PendulumNonInteractiveEnv — Type

A non-interactive pendulum environment.

Accepts only nothing actions, which result in the system being simulated for one time step. Sets env.done to true once maximum_time is reached. Resets to a random position and momentum. Always returns zero rewards.

Useful for debugging and development purposes, particularly in model-based reinforcement learning.

ReinforcementLearningEnvironments.PendulumNonInteractiveEnv — Method

PendulumNonInteractiveEnv(;kwargs...)

Keyword arguments

float_type = Float64

gravity = 9.8

length = 2.0

mass = 1.0

step_size = 0.01

maximum_time = 10.0

rng = Random.GLOBAL_RNG