ReinforcementLearningZoo.jl

ReinforcementLearningZoo.A2CGAELearner — Type

A2CGAELearner(;kwargs...)

Keyword arguments

approximator, an ActorCritic based NeuralNetworkApproximator

γ::Float32, reward discount rate.

'λ::Float32', lambda for GAE-lambda

actor_loss_weight::Float32

critic_loss_weight::Float32

entropy_loss_weight::Float32

ReinforcementLearningZoo.A2CLearner — Type

A2CLearner(;kwargs...)

Keyword arguments

approximator::ActorCritic

γ::Float32, reward discount rate.

actor_loss_weight::Float32

critic_loss_weight::Float32

entropy_loss_weight::Float32

ReinforcementLearningZoo.BasicDQNLearner — Type

BasicDQNLearner(;kwargs...)

See paper: Playing Atari with Deep Reinforcement Learning

This is the very basic implementation of DQN. Compared to the traditional Q learning, the only difference is that, in the updating step it uses a batch of transitions sampled from an experience buffer instead of current transition. And the approximator is usually a NeuralNetworkApproximator. You can start from this implementation to understand how everything is organized and how to write your own customized algorithm.

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

loss_func: the loss function to use. TODO: provide a default huber_loss?

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

rng=Random.GLOBAL_RNG

ReinforcementLearningZoo.BestResponsePolicy — Method

BestResponsePolicy(policy, env, best_responder)

policy, the original policy to be wrapped in the best response policy.

env, the environment to handle.

best_responder, the player to choose best response action.

ReinforcementLearningZoo.DDPGPolicy — Method

DDPGPolicy(;kwargs...)

Keyword arguments

behavior_actor,

behavior_critic,

target_actor,

target_critic,

start_policy,

γ = 0.99f0,

ρ = 0.995f0,

batch_size = 32,

start_steps = 10000,

update_after = 1000,

update_every = 50,

act_limit = 1.0,

act_noise = 0.1,

step = 0,

rng = Random.GLOBAL_RNG,

ReinforcementLearningZoo.DQNLearner — Type

DQNLearner(;kwargs...)

See paper: Human-level control through deep reinforcement learning

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

rng = Random.GLOBAL_RNG

ReinforcementLearningZoo.DQNLearner — Method

Note

The state of the observation is assumed to have been stacked, if !isnothing(stack_size).

ReinforcementLearningZoo.DeepCFR — Type

DeepCFR(;kwargs...)

Symbols used here follow the paper: Deep Counterfactual Regret Minimization

Keyword arguments

K, number of traverrsal.

t, number of iteration.

Π, the policy network.

V, a dictionary of each player's advantage network.

MΠ, a strategy memory.

MV, a dictionary of each player's advantage memory.

reinitialize_freq=1, the frequency of reinitializing the value networks.

ReinforcementLearningZoo.ExternalSamplingMCCFRPolicy — Type

ExternalSamplingMCCFRPolicy

This implementation uses stochasticaly-weighted averaging.

Ref:

MONTE CARLO SAMPLING AND REGRET MINIMIZATION FOR EQUILIBRIUM COMPUTATION AND DECISION-MAKING IN LARGE EXTENSIVE FORM GAMES

Monte Carlo Sampling for Regret Minimization in Extensive Games

ReinforcementLearningZoo.GaussianNetwork — Type

GaussianNetwork(;pre=identity, μ, σ)

σ should return the log of std, exp will be applied to it automatically.

ReinforcementLearningZoo.IQNLearner — Type

IQNLearner(;kwargs)

See paper

Keyworkd arugments

approximator, a ImplicitQuantileNet

target_approximator, a ImplicitQuantileNet, must have the same structure as approximator

κ = 1.0f0,

N = 32,

N′ = 32,

Nₑₘ = 64,

K = 32,

γ = 0.99f0,

stack_size = 4,

batch_size = 32,

update_horizon = 1,

min_replay_history = 20000,

update_freq = 4,

target_update_freq = 8000,

update_step = 0,

default_priority = 1.0f2,

β_priority = 0.5f0,

rng = Random.GLOBAL_RNG,

device_seed = nothing,

ReinforcementLearningZoo.ImplicitQuantileNet — Type

ImplicitQuantileNet(;ψ, ϕ, header)

        quantiles (n_action, n_quantiles, batch_size)
           ↑
         header
           ↑
feature ↱  ⨀   ↰ transformed embedding
       ψ       ϕ
       ↑       ↑
       s        τ

ReinforcementLearningZoo.MinimaxPolicy — Type

MinimaxPolicy(;value_function, depth::Int)

The minimax algorithm with Alpha-beta pruning

Keyword Arguments

maximum_depth::Int=30, the maximum depth of search.

value_function=nothing, estimate the value of env. value_function(env) -> Number. It is only called after searching for maximum_depth and the env is not terminated yet.

ReinforcementLearningZoo.OutcomeSamplingMCCFRPolicy — Type

OutcomeSamplingMCCFRPolicy

This implementation uses stochasticaly-weighted averaging.

Ref:

MONTE CARLO SAMPLING AND REGRET MINIMIZATION FOR EQUILIBRIUM COMPUTATION AND DECISION-MAKING IN LARGE EXTENSIVE FORM GAMES

Monte Carlo Sampling for Regret Minimization in Extensive Games

ReinforcementLearningZoo.PPOPolicy — Type

PPOPolicy(;kwargs)

Keyword arguments

approximator,

γ = 0.99f0,

λ = 0.95f0,

clip_range = 0.2f0,

max_grad_norm = 0.5f0,

n_microbatches = 4,

n_epochs = 4,

actor_loss_weight = 1.0f0,

critic_loss_weight = 0.5f0,

entropy_loss_weight = 0.01f0,

dist = Categorical,

rng = Random.GLOBAL_RNG,

By default, dist is set to Categorical, which means it will only works on environments of discrete actions. To work with environments of

ReinforcementLearningZoo.PrioritizedDQNLearner — Type

PrioritizedDQNLearner(;kwargs...)

See paper: Prioritized Experience Replay And also https://danieltakeshi.github.io/2019/07/14/per/

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

rng = Random.GLOBAL_RNG

ReinforcementLearningZoo.PrioritizedDQNLearner — Method

Note

The state of the observation is assumed to have been stacked, if !isnothing(stack_size).

ReinforcementLearningZoo.RainbowLearner — Type

RainbowLearner(;kwargs...)

See paper: Rainbow: Combining Improvements in Deep Reinforcement Learning

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

Vₘₐₓ::Float32: the maximum value of distribution.

Vₘᵢₙ::Float32: the minimum value of distribution.

n_actions::Int: number of possible actions.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=500: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float32=1.0f2.: the default priority for newly added transitions. It must be >= 1.

n_atoms::Int=51: the number of buckets of the value function distribution.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

rng = Random.GLOBAL_RNG

ReinforcementLearningZoo.SACPolicy — Method

SACPolicy(;kwargs...)

Keyword arguments

policy,

qnetwork1,

qnetwork2,

target_qnetwork1,

target_qnetwork2,

start_policy,

γ = 0.99f0,

ρ = 0.995f0,

α = 0.2f0,

batch_size = 32,

start_steps = 10000,

update_after = 1000,

update_every = 50,

step = 0,

rng = Random.GLOBAL_RNG,

ReinforcementLearningZoo.TD3Policy — Method

TD3Policy(;kwargs...)

Keyword arguments

behavior_actor,

behavior_critic,

target_actor,

target_critic,

start_policy,

γ = 0.99f0,

ρ = 0.995f0,

batch_size = 32,

start_steps = 10000,

update_after = 1000,

update_every = 50,

policy_freq = 2 # frequency in which the actor performs a gradient step and critic target is updated

target_act_limit = 1.0, # noise added to actor target

target_act_noise = 0.1, # noise added to actor target

act_limit = 1.0, # noise added when outputing action

act_noise = 0.1, # noise added when outputing action

step = 0,

rng = Random.GLOBAL_RNG,

ReinforcementLearningZoo.TabularCFRPolicy — Method

TabularCFRPolicy(;kwargs...)

Some useful papers while implementing this algorithm:

An Introduction to Counterfactual Regret Minimization

MONTE CARLO SAMPLING AND REGRET MINIMIZATION FOR EQUILIBRIUM COMPUTATION AND DECISION-MAKING IN LARGE EXTENSIVE FORM GAMES

Solving Large Imperfect Information Games Using CFR⁺

Revisiting CFR⁺ and Alternating Updates

Solving Imperfect-Information Games via Discounted Regret Minimization

Keyword Arguments

is_alternating_update=true: If true, we update the players alternatively.

is_reset_neg_regrets=true: Whether to use regret matching⁺.

is_linear_averaging=true

weighted_averaging_delay=0. The averaging delay in number of iterations. Only valid when is_linear_averaging is set to true.

state_type=String, the data type of information set.

rng=Random.GLOBAL_RNG

ReinforcementLearningZoo.VPGPolicy — Type

Vanilla Policy Gradient

VPGPolicy(;kwargs)

Keyword arguments

approximator,

baseline,

dist, distribution function of the action

γ, discount factor

α_θ, step size of policy parameter

α_w, step size of baseline parameter

batch_size,

rng,

loss,

baseline_loss,

if the action space is continuous, then the env should transform the action value, (such as using tanh), in order to make sure low ≤ value ≤ high

ReinforcementLearningBase.update! — Method

Run one interation

ReinforcementLearningBase.update! — Method

Update Π (policy network)

ReinforcementLearningBase.update! — Method

Run one interation

ReinforcementLearningBase.update! — Method

Run one interation

ReinforcementLearningBase.update! — Method

Run one interation

ReinforcementLearningBase.update! — Method

Update the behavior_policy

ReinforcementLearningZoo.cfr! — Function

Symbol meanings:

π: reach prob π′: new reach prob π₋ᵢ: opponents' reach prob p: player to update. nothing means simultaneous update. w: weight v: counterfactual value before weighted by opponent's reaching probability V: a vector containing the v after taking each action with current information set. Used to calculate the regret value

ReinforcementLearningZoo.evaluate — Method

This function is compatible with a multidimensional action space.

ReinforcementLearningZoo.external_sampling! — Method

CFR Traversal with External Sampling

ReinforcementLearningZoo.masked_regret_matching — Method

This is the specific regret matching method used in DeepCFR

ReinforcementLearningZoo.update_advantage_networks — Method

Update advantage network