Utils

ReinforcementLearning.Utils.SumTree — Type

SumTree(capacity::Int)

Efficiently sample and update weights. For more detals, see the post at here. Here we use a vector to represent the binary tree. Suppose we will have capacity leaves at most. Every time we push! new node into the tree, only the recent capacity node and their sum will be updated! [––––––Parent nodes––––––][––––leaves––––] [size: 2^ceil(Int, log2(capacity))-1 ][ size: capacity ]

Example

julia> t = SumTree(8)
0-element SumTree
julia> for i in 1:16
       push!(t, i)
       end
julia> t
8-element SumTree:
  9.0
 10.0
 11.0
 12.0
 13.0
 14.0
 15.0
 16.0
julia> sample(t)
(2, 10.0)
julia> sample(t)
(1, 9.0)
julia> inds, ps = sample(t,100000)
([8, 4, 8, 1, 5, 2, 2, 7, 6, 6  …  1, 1, 7, 1, 6, 1, 5, 7, 2, 7], [16.0, 12.0, 16.0, 9.0, 13.0, 10.0, 10.0, 15.0, 14.0, 14.0  …  9.0, 9.0, 15.0, 9.0, 14.0, 9.0, 13.0, 15.0, 10.0, 15.0])
julia> countmap(inds)
Dict{Int64,Int64} with 8 entries:
  7 => 14991
  4 => 12019
  2 => 10003
  3 => 11027
  5 => 12971
  8 => 16052
  6 => 13952
  1 => 8985
julia> countmap(ps)
Dict{Float64,Int64} with 8 entries:
  9.0  => 8985
  13.0 => 12971
  10.0 => 10003
  14.0 => 13952
  16.0 => 16052
  11.0 => 11027
  15.0 => 14991
  12.0 => 12019

ReinforcementLearning.Utils.Tiling — Type

Tiling(ranges::NTuple{N, Tr}) where {N, Tr}

Using a tuple of ranges to simulate a tiling. The length of ranges indicates the dimension of tilling.

Example

julia> t = Tiling((1:2:5, 10:5:20))
Tiling{2,StepRange{Int64,Int64}}((1:2:5, 10:5:20), [1 3; 2 4])

julia> encode(t, (2, 12))  # encode into an Int
1

julia> encode(t, (2, 18))
3

julia> t2 = t - (1, 3)  # shift a little to get a new Tiling
Tiling{2,StepRange{Int64,Int64}}((0:2:4, 7:5:17), [1 3; 2 4])

ReinforcementLearning.Utils.findallmax — Method

findallmax(A::AbstractArray)

Like findmax, but all the indices of the maximum value are returned.

Warning

All elements of value NaN in A will be ignored, unless all elements are NaN. In that case, the returned maximum value will be NaN and the returned indices will be collect(1:length(A))

#Examples

julia> findallmax([-Inf, -Inf, -Inf])
(-Inf, [1, 2, 3])

julia> findallmax([Inf, Inf, Inf])
(Inf, [1, 2, 3])

julia> findallmax([Inf, 0, Inf])
(Inf, [1, 3])

julia> findallmax([0,1,2,1,2,1,0])
(2, [3, 5])

ReinforcementLearning.Utils.huber_loss — Method

huber_loss(labels, predictions;δ = 1.0)

See huber_loss and the implementation in TensorFlow.

Warning

The return is not reduced!

Base.:- — Method

(-)(t::Tiling, xs)

Shift t along each dimension by each element in xs.

ReinforcementLearning.Utils.get — Method

!!! this is unsafe, always check the real_ind, or you may get bound error in some extreme cases.