Approximators

Approximators

ReinforcementLearning.AbstractApproximator — Type.

An approximator is a functional object to estimate a state. Two typical kinds of approximators are AbstractVApproximator and AbstractQApproximator.

source

ReinforcementLearning.AbstractVApproximator — Type.

A collection of approximators to estimate the value of a state.

source

ReinforcementLearning.AbstractQApproximator — Type.

A collection of approximators to estimate the values of actions given a state.

source

ReinforcementLearning.LinearVApproximator — Type.

LinearVApproximator(weights::Array{Float64, N}) -> LinearVApproximator{N}

Use the weighted sum to represent the estimation of a state. The state is expected to have the same length with weights.

See also LinearQApproximator

source

ReinforcementLearning.TabularVApproximator — Type.

TabularVApproximator(table) -> TabularVApproximator
TabularVApproximator(ns::Int, init::Float64=0.0) -> TabularVApproximator

Use a table of type Vector{Float64} of length ns to record the state values.

source

ReinforcementLearning.LinearQApproximator — Type.

LinearQApproximator(weights::Vector{Float64}, feature_func::F, actions::Vector{Int}) -> LinearQApproximator{F}

Use weighted sum to represent the estimation given a state and an action.

Fields

weights::Vector{Float64}: the weight of each feature.

feature_func::Function: decide how to generate a feature vector of length(weights) given a state and an action as parameters.

actions::Vector{Int}: all possible actions.

See also LinearVApproximator.

source

ReinforcementLearning.NeuralNetworkQ — Type.

NeuralNetworkQ(;kwargs...) -> NeuralNetworkQ{D, M, O, P}

Use neural networks to generate estimations of state-action values.

Keywords

model::M: describes the network structure.

optimizer::O: defines how to update parameters given grads.

parameters::P=params(model): the parameters of model.

device::Symbol=:cpu: the param D of NeuralNetworkQ, specify where to run the model. Supported keywords are:

:cpu

:gpu, if the specified device is :gpu, then we will automatically change the device to one of the following device types according to the backend(model):

:Knet_gpu, means the model is Knet that runs on gpu.

:Zygote_gpu, means the model is Flux with Zygote that runs on gpu.

Fields

model::M

optimizer::O

params::P

source