Learners

Learners

ReinforcementLearning.AbstractLearner — Type.

A learner is used to define:

How to generate necessary training data?

How to update the inner Approximators?

source

Traditional Learners

ReinforcementLearning.TDLearner — Type.

TDLearner(approximator::Tapp, γ::Float64, optimizer::Float64; n::Int=0) where {Tapp<:AbstractVApproximator}
TDLearner(approximator::Tapp, γ::Float64, optimizer::Float64; n::Int=0, method::Symbol=:SARSA) where {Tapp<:AbstractQApproximator}

The TDLearner(Temporal Difference Learner) use the latest n step experiences to update the approximator. Note that n starts with 0, which means looking forward for the next n steps. γ is the discount rate of experience. optimizer is the learning rate.

For AbstractVApproximator, the only supported update method is :SRS, which means only States, Rewards and next_Sates are used to update the approximator.

For AbstractQApproximator, the following methods are supported:

:SARS (aka Q-Learning)

:SARSA

:ExpectedSARSA

source

ReinforcementLearning.DoubleLearner — Type.

DoubleLearner(;L1, L2)

For now, this only supports TDLearner for L1 and L2.

source

ReinforcementLearning.DifferentialTDLearner — Type.

DifferentialTDLearner(;approximator::A, α::Float64, β::Float64, R̄::Float64 = 0.0, n::Int = 0)

source

ReinforcementLearning.TDλReturnLearner — Type.

TDλReturnLearner(;approximator::Tapp, γ::Float64 = 1.0, α::Float64, λ::Float64)

source

ReinforcementLearning.MonteCarloLearner — Type.

MonteCarloLearner(; approximator::A, γ = 1.0, α = 1.0, kind = FIRST_VISIT, sampling = NO_SAMPLING, returns = CachedSampleAvg())

source

ReinforcementLearning.GradientBanditLearner — Type.

GradientBanditLearner(;approximator::A, optimizer::O, baseline::B)

source

Q Learners

ReinforcementLearning.BasicDQNLearner — Type.

BasicDQNLearner(;kwargs...)

See paper: Playing Atari with Deep Reinforcement Learning

This is the very basic implementation of DQN. Compared to the traditional Q learning, the only difference is that, in the updating step it uses a batch of transitions sampled from an experience buffer instead of current transition. And the approximator is usually a NeuralNetworkQ.

You can start from this implementation to understand how everything is organized and how to write your own customized algorithm.

Keywords

approximator::AbstractQApproximator: used to get Q-values of a state.

loss_fun: the loss function to use. TODO: provide a default huber_loss?

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

source

ReinforcementLearning.DQNLearner — Type.

DQNLearner(;kwargs...)

See paper: Human-level control through deep reinforcement learning

Keywords

approximator::AbstractQApproximator: used to get Q-values of a state.

target_approximator::AbstractQApproximator: similar to approximator, but used to estimate the target (the next state).

loss_fun: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

source

ReinforcementLearning.PrioritizedDQNLearner — Type.

PrioritizedDQNLearner(;kwargs...)

See paper: Prioritized Experience Replay

Keywords

approximator::AbstractQApproximator: used to get Q-values of a state.

target_approximator::AbstractQApproximator: similar to approximator, but used to estimate the target (the next state).

loss_fun: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

source

ReinforcementLearning.RainbowLearner — Type.

RainbowLearner(;kwargs...)

See paper: Rainbow: Combining Improvements in Deep Reinforcement Learning

Keywords

approximator::AbstractQApproximator: used to get Q-values of a state.

target_approximator::AbstractQApproximator: similar to approximator, but used to estimate the target (the next state).

loss_fun: the loss function.

Vₘₐₓ::Float32: the maximum value of distribution.

Vₘᵢₙ::Float32: the minimum value of distribution.

n_actions::Int: number of possible actions.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=500: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

n_atoms::Int=51: the number of buckets of the value function distribution.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

Warning

The Zygote_gpu version is slow due to that the argmax(A, dims=1) falls back to the CPU version in CuArrays.

source