Learners

Learners

ReinforcementLearning.AbstractLearner — Type.

A learner is used to define:

How to generate necessary training data?

How to update the inner Approximators?

source

Traditional Learners

ReinforcementLearning.TDLearner — Type.

TDLearner(approximator::Tapp, γ::Float64, optimizer::Float64; n::Int=0) where {Tapp<:AbstractVApproximator}
TDLearner(approximator::Tapp, γ::Float64, optimizer::Float64; n::Int=0, method::Symbol=:SARSA) where {Tapp<:AbstractQApproximator}

The TDLearner(Temporal Difference Learner) use the latest n step experiences to update the approximator. Note that n starts with 0, which means looking forward for the next n steps. γ is the discount rate of experience. optimizer is the learning rate.

For AbstractVApproximator, the only supported update method is :SRS, which means only States, Rewards and next_Sates are used to update the approximator.

For AbstractQApproximator, the following methods are supported:

:SARS (aka Q-Learning)

:SARSA

:ExpectedSARSA

source

ReinforcementLearning.DoubleLearner — Type.

DoubleLearner(;L1, L2)

For now, this only supports TDLearner for L1 and L2.

source

ReinforcementLearning.DifferentialTDLearner — Type.

DifferentialTDLearner(;approximator::A, α::Float64, β::Float64, R̄::Float64 = 0.0, n::Int = 0)

source

ReinforcementLearning.TDλReturnLearner — Type.

TDλReturnLearner(;approximator::Tapp, γ::Float64 = 1.0, α::Float64, λ::Float64)

source

ReinforcementLearning.MonteCarloLearner — Type.

MonteCarloLearner(; approximator::A, γ = 1.0, α = 1.0, kind = FIRST_VISIT, sampling = NO_SAMPLING, returns = CachedSampleAvg())

source

ReinforcementLearning.GradientBanditLearner — Type.

GradientBanditLearner(;approximator::A, optimizer::O, baseline::B)

source

Q Learners

ReinforcementLearning.BasicDQNLearner — Type.

BasicDQNLearner(;kwargs...)

See paper: Playing Atari with Deep Reinforcement Learning

Keywords

TODO: need review

source

ReinforcementLearning.DQNLearner — Type.

DQNLearner(;kwargs...)

See paper: Human-level control through deep reinforcement learning

Keywords

TODO: need review

source

ReinforcementLearning.PrioritizedDQNLearner — Type.

PrioritizedDQNLearner(;kwargs...)

See paper: Prioritized Experience Replay

Keywords

TODO: need review

source

ReinforcementLearning.RainbowLearner — Type.

RainbowLearner(;kwargs...)

See paper: Rainbow: Combining Improvements in Deep Reinforcement Learning

Keywords

TODO: need review

source