Policies

Policies

ReinforcementLearning.AbstractPolicy — Type.

A policy is a functional object to generate an action given a state.

source

ReinforcementLearning.TabularRandomPolicy — Type.

TabularRandomPolicy(prob::Array{Float64, 2})

prob describes the distribution of actions for each state.

source

ReinforcementLearning.TabularDeterministicPolicy — Type.

TabularDeterministicPolicy(table::Vector{Int}, nactions::Int) -> TabularDeterministicPolicy

Always select the specific action according to the state defined in table. nactions is used to define the number of possible actions.

source

ReinforcementLearning.ExploringStartPolicy — Type.

ExploringStartPolicy(π::P, actions::A) -> ExploringStartPolicy{P,A}

Generate a random action at the beginning of an episode. For other steps, use π to generate the action.

source

ReinforcementLearning.OffPolicy — Type.

OffPolicy(π_target::P, π_behavior::B) -> OffPolicy{P,B}

source

ReinforcementLearning.QBasedPolicy — Type.

QBasedPolicy(learner::Q, selector::S) -> QBasedPolicy{Q, S}

Use a Q-learner to generate the estimations of actions and use selector to get the action.

source

ReinforcementLearning.VBasedPolicy — Type.

VBasedPolicy(learner::V, f::F) -> VBasedPolicy{V,F}

Use a function f to generate action and update the V-learner in the meantime.

source

ReinforcementLearning.ReinforcePolicy — Type.

ReinforcePolicy(approximator, α, γ)

This is the very basic reinforce algorithm.

TODO: implement some other variants.

source