Learners

Learners

TD Learner

ReinforcementLearning.ExpectedSarsa — Method.

ExpectedSarsa(; kargs...) = TDLearner(; endvaluepolicy = ExpectedSarsaEndPolicy(VeryOptimisticEpsilonGreedyPolicy(.1)), kargs...)

source

ReinforcementLearning.QLearning — Method.

QLearning(; kargs...) = TDLearner(; endvaluepolicy = QLearningEndPolicy(), kargs...)

source

ReinforcementLearning.Sarsa — Method.

Sarsa(; kargs...) = TDLearner(; kargs...)

source

ReinforcementLearning.AccumulatingTraces — Type.

struct AccumulatingTraces <: AbstractTraces
    λ::Float64
    γλ::Float64
    trace::Array{Float64, 2}
    minimaltracevalue::Float64

Decaying traces with factor γλ.

Traces are updated according to $e(a, s) ← 1 + e(a, s)$ for the current action-state pair and $e(a, s) ← γλ e(a, s)$ for all other pairs unless $e(a, s) <$ minimaltracevalue where the trace is set to 0 (for computational efficiency).

source

ReinforcementLearning.AccumulatingTraces — Method.

AccumulatingTraces(ns, na, λ::Float64, γ::Float64; minimaltracevalue = 1e-12)

source

ReinforcementLearning.NoTraces — Type.

struct NoTraces <: AbstractTraces

No eligibility traces, i.e. $e(a, s) = 1$ for current action $a$ and state $s$ and zero otherwise.

source

ReinforcementLearning.ReplacingTraces — Type.

struct ReplacingTraces <: AbstractTraces
    λ::Float64
    γλ::Float64
    trace::Array{Float64, 2}
    minimaltracevalue::Float64

Decaying traces with factor γλ.

Traces are updated according to $e(a, s) ← 1$ for the current action-state pair and $e(a, s) ← γλ e(a, s)$ for all other pairs unless $e(a, s) <$ minimaltracevalue where the trace is set to 0 (for computational efficiency).

source

ReinforcementLearning.ReplacingTraces — Method.

ReplacingTraces(ns, na, λ::Float64, γ::Float64; minimaltracevalue = 1e-12)

source

Initial values, novel actions and unseen values

For td-error dependent methods, the exploration-exploitation trade-off depends on the initvalue and the unseenvalue. To distinguish actions that were never choosen before, i.e. novel actions, the default initial Q-value (field param) is initvalue = Inf64. In a state with novel actions, the policy determines how to deal with novel actions. To compute the td-error the unseenvalue is used for states with novel actions. One way to achieve agressively exploratory behavior is to assure that unseenvalue (or initvalue) is larger than the largest possible Q-value.

Policy Gradient Learner

ReinforcementLearning.Critic — Type.

mutable struct Critic <: AbstractBiasCorrector
    α::Float64
    V::Array{Float64, 1}

source

ReinforcementLearning.Critic — Method.

Critic(; γ = .9, α = .1, ns = 10, initvalue = 0.)

source

ReinforcementLearning.NoBiasCorrector — Type.

struct NoBiasCorrector <: AbstractBiasCorrector

source

ReinforcementLearning.RewardLowpassFilterBiasCorrector — Type.

mutable struct RewardLowpassFilterBiasCorrector <: AbstractBiasCorrector
λ::Float64
rmean::Float64

Filters the reward with factor λ and uses effective reward (r - rmean) to update the parameters.

source

ReinforcementLearning.ActorCriticPolicyGradient — Method.

ActorCriticPolicyGradient(; nsteps = 1, γ = .9, ns = 10, na = 4, 
                            α = .1, αcritic = .1, initvalue = Inf64)

source

ReinforcementLearning.EpisodicReinforce — Method.

EpisodicReinforce(; kwargs...) = PolicyGradientForward(; kwargs...)

source

ReinforcementLearning.AbstractPolicyGradient — Type.

mutable struct PolicyGradientBackward <: AbstractPolicyGradient
    ns::Int64 = 10
    na::Int64 = 4
    γ::Float64 = .9
    α::Float64 = .1
    initvalue::Float64 = 0.
    params::Array{Float64, 2} = zeros(na, ns) + initvalue
    traces::AccumulatingTraces = AccumulatingTraces(ns, na, 1., γ, 
                                                    trace = zeros(na, ns))
    biascorrector::T = NoBiasCorrector()

Policy gradient learning in the backward view.

The parameters are updated according to $params[a, s] += α * r_{eff} * e[a, s]$ where $r_{eff} = r$ for NoBiasCorrector, $r_{eff} = r - rmean$ for RewardLowpassFilterBiasCorrector and e[a, s] is the eligibility trace.

source

N-step Learner

Model Based Learner

ReinforcementLearning.policy_iteration! — Method.

policy_iteration!(mdplearner::MDPLearner)

Solve MDP with policy iteration using MDPLearner.

source

Deep Reinforcement Learning