ReinforcementLearningBase.jl

ReinforcementLearningBase.RLBase — Module

ReinforcementLearningBase.jl (RLBase) provides some common constants, traits, abstractions and interfaces in developing reinforcement learning algorithms in Julia. From the concept level, they can be organized in the following parts:

General

Agent

Policy

Learner

Approximator

Explorer

EnvironmentModel

Environment

Preprocessor

Observation

Trajectory

Space

General

ReinforcementLearningBase.AbstractStage — Type

In the simplest case, we just need to run env |> observe |> agent |> env repeatly. But usually we would also like to control when/how to update the agent. So we defined the following stages for the help:

PRE_EXPERIMENT_STAGE

PRE_EPISODE_STAGE

PRE_ACT_STAGE

POST_ACT_STAGE

POST_EPISODE_STAGE

POST_EXPERIMENT_STAGE

                      +-----------------------------------------------------------+                      
                      |Episode                                                    |                      
                      |                                                           |                      
PRE_EXPERIMENT_STAGE  |            PRE_ACT_STAGE    POST_ACT_STAGE                | POST_EXPERIMENT_STAGE
         |            |                  |                |                       |          |           
         v            |        +-----+   v   +-------+    v   +-----+             |          v           
         --------------------->+ env +------>+ agent +------->+ env +---> ... ------->......             
                      |  ^     +-----+  obs  +-------+ action +-----+          ^  |                      
                      |  |                                                     |  |                      
                      |  +--PRE_EPISODE_STAGE            POST_EPISODE_STAGE----+  |                      
                      |                                                           |                      
                      |                                                           |                      
                      +-----------------------------------------------------------+                      

Agent

ReinforcementLearningBase.AbstractAgent — Type

(agent::AbstractAgent)(obs) = agent(PRE_ACT_STAGE, obs) -> action
(agent::AbstractAgent)(stage::AbstractStage, obs)

An agent is a functional object which takes in an observation and returns an action. In different stages, the behavior may be different.

ReinforcementLearningBase.get_role — Function

return DEFAULT_PLAYER by default

ReinforcementLearningBase.DEFAULT_PLAYER — Constant

Default player instance for all the environments

Policy

ReinforcementLearningBase.AbstractPolicy — Type

(p::AbstractPolicy)(obs) = p(obs, ActionStyle(obs)) -> action

Similar to AbstractAgent, a policy is aslo a functional object which defines how to generate action(s) given an observation of the environment. However, in the concept level, a policy is more lower and usually it doesn't need to care about how/when to interact with an environment.

ReinforcementLearningBase.update! — Method

update!(p::AbstractPolicy, experience)

Update the policy p with experience

ReinforcementLearningBase.get_prob — Method

get_prob(p::AbstractPolicy, obs) = get_prob(p, obs, ActionStyle(obs))
get_prob(p::AbstractPolicy, obs, ::AbstractActionStyle)

Get the probability distribution of actions from the policy p given an observation obs

ReinforcementLearningBase.get_prob — Method

get_prob(p::AbstractPolicy, obs, a) = get_prob(p, obs, ActionStyle(obs), a)

Get the probability of to take action a from the policy p given an observation obs

ReinforcementLearningBase.get_priority — Function

get_priority(p::AbstractLearner, experience)

get_priority(p::AbstractPolicy, experience)

Calculate the priority of the experience to policy p

Learner

ReinforcementLearningBase.AbstractLearner — Type

(learner::AbstractLearner)(obs)

A learner is usually a wrapper around AbstractApproximators. It defines the expected inputs and how to update inner approximators. From the concept level, it assumes that the necessary training data is already cooked (usually by AbstractPolicy with the extract_experience method).

ReinforcementLearningBase.update! — Method

update!(learner::AbstractLearner, experience)

ReinforcementLearningBase.get_priority — Method

get_priority(p::AbstractLearner, experience)

ReinforcementLearningBase.extract_experience — Method

extract_experience(trajectory::AbstractTrajectory, learner::AbstractLearner)

Extract necessary data from the trajectory given a learner

Approximator

ReinforcementLearningBase.AbstractApproximator — Type

(app::AbstractApproximator)(obs) = app(get_state(obs))

An approximator is a functional object for value estimation. It serves as a black box to provides an abstraction over different kinds of approximate methods (for example DNN provided by Flux or Knet).

ReinforcementLearningBase.batch_estimate — Function

batch_estimate(app::AbstractApproximator, states)

The states is assume to be a batch of states

ReinforcementLearningBase.update! — Method

update!(a::AbstractApproximator, correction)

Usually the correction is the gradient of inner parameters.

Base.copyto! — Method

Base.copyto!(dest::AbstractApproximator, src::AbstractApproximator)

Copy the internal params from src approximator to dest approximator

ReinforcementLearningBase.ApproximatorStyle — Function

ApproximatorStyle(x::AbstractApproximator)

Detect which kind of approximator it is

ReinforcementLearningBase.Q_APPROXIMATOR — Constant

For Q_APPROXIMATOR, we assume that the following methods are implemented:

(Q::AbstractApproximator)(s, a), estimate the Q value.

(Q::AbstractApproximator)(s), estimate the Q value among all possible actions.

ReinforcementLearningBase.V_APPROXIMATOR — Constant

For V_APPROXIMATOR, we assume that (V::AbstractApproximator)(s) is implemented.

ReinforcementLearningBase.HYBRID_APPROXIMATOR — Constant

For HYBRID_APPROXIMATOR, the following methods are assumed to be implemented:

(app::AbstractApproximator)(s, ::Val{:Q}), estimate the action values of state s.

(app::AbstractApproximator)(s, ::Val{:V}), estimate the state values.

(app::AbstractApproximator)(s, a), estimate state-action values.

Explorer

ReinforcementLearningBase.AbstractExplorer — Type

(p::AbstractExplorer)(x)
(p::AbstractExplorer)(x, mask)

Define how to select an action based on action values.

ReinforcementLearningBase.reset! — Method

Reset the internal state of an explorer p

Base.copy — Method

The internal state is also copied

ReinforcementLearningBase.get_prob — Method

get_prob(p::AbstractExplorer, x) -> AbstractDistribution

Get the action distribution given action values

ReinforcementLearningBase.get_prob — Method

get_prob(p::AbstractExplorer, x, mask)

Similart to get_prob(p::AbstractExplorer, x), but here only the masked elements are considered.

EnvironmentModel

ReinforcementLearningBase.AbstractEnvironmentModel — Type

Describe how to model a reinforcement learning environment.

TODO: need more investigation

Ref: https://bair.berkeley.edu/blog/2019/12/12/mbpo/

Analytic gradient computation

Sampling-based planning

Model-based data generation

Value-equivalence prediction

Environment

ReinforcementLearningBase.AbstractEnv — Type

(env::AbstractEnv)(action) = env(DEFAULT_PLAYER, action) -> nothing
(env::AbstractEnv)(player, action) -> nothing

Super type of all reinforcement learning environments. An environments is a functional object which takes in an action and returns nothing.

Note

So why don't we adopt the step! method like OpenAI Gym here? The reason is that the async manner will simplify a lot of things here.

ReinforcementLearningBase.DynamicStyle — Function

DynamicStyle(x::AbstractEnv) = SEQUENTIAL

Determine whether the players can play simultaneous or not. Default value is SEQUENTIAL

ReinforcementLearningBase.SIMULTANEOUS — Constant

Environment with the DynamicStyle of SIMULTANEOUS must take in actions from some (or all) players at one time

ReinforcementLearningBase.SEQUENTIAL — Constant

Environment with the DynamicStyle of SEQUENTIAL must takes actions from different players one-by-one.

ReinforcementLearningBase.get_action_space — Function

get_action_space(env::AbstractEnv) -> AbstractSpace

ReinforcementLearningBase.get_observation_space — Function

get_observation_space(env::AbstractEnv) -> AbstractSpace

ReinforcementLearningBase.get_current_player — Function

Return DEFAULT_PLAYER by default

ReinforcementLearningBase.get_num_players — Function

get_num_players(env::AbstractEnv) -> Int

ReinforcementLearningBase.render — Function

Show the environment in a user-friendly manner

ReinforcementLearningBase.reset! — Method

Reset the internal state of an environment

Several meta-environments are also provided:

ReinforcementLearningBase.WrappedEnv — Type

WrappedEnv(;preprocessor, env, postprocessor=identity)

The observation of the inner env is first transformed by the preprocessor. And the action is transformed by postprocessor and then send to the inner env.

Preprocessor

ReinforcementLearningBase.AbstractPreprocessor — Type

Preprocess an observation and return a new observation.

Observation

ReinforcementLearningBase.observe — Function

observe(env::AbstractEnv) = observe(env, get_current_player(env))
observe(::AbstractEnv, player)

Get an observation of the env from the perspective of an player.

Note

This is a very deliberate decision to adopt the duck-typing here to describe an observation from an environment. By default, we assume an observation is a NamedTuple, which is the most common case. But of course it can be of any type, as long as it implemented the necessay methods described in this section.

ReinforcementLearningBase.get_legal_actions_mask — Function

get_legal_actions_mask(x) -> Bool[]

Only valid for observations of FULL_ACTION_SET.

ReinforcementLearningBase.get_legal_actions — Function

get_legal_actions(x)

Only valid for observations of FULL_ACTION_SET.

ReinforcementLearningBase.get_terminal — Function

get_terminal(x) -> bool

ReinforcementLearningBase.get_reward — Function

get_reward(x) -> Number

ReinforcementLearningBase.get_state — Function

get_state(x) -> Array

ReinforcementLearningBase.get_invalid_action — Function

By default INVALID_ACTION is returned

ReinforcementLearningBase.INVALID_ACTION — Constant

The default value of invalid action for all the environments. However, one can still specify the value for a specific environment or observation.

See also: get_invalid_action

ReinforcementLearningBase.ActionStyle — Function

ActionStyle(x)

Specify whether the observation contains a full action set or a minimal action set. By default the MINIMAL_ACTION_SET is returned.

ReinforcementLearningBase.MINIMAL_ACTION_SET — Constant

All actions in the action space of the environment are legal

ReinforcementLearningBase.FULL_ACTION_SET — Constant

The action space of the environment may contains illegal actions

ReinforcementLearningBase.RTSA — Constant

An alias of (:reward, :terminal, :state, :action)

ReinforcementLearningBase.SARTSA — Constant

An alias of (:state, :action, :reward, :terminal, :next_state, :next_action)

Some meta-observations are also provided:

ReinforcementLearningBase.StateOverriddenObs — Type

StateOverriddenObs(;obs, state)

Replace the internal state of obs with state.

Example

julia> old_obs = (reward=1.0, terminal=false, state=1)
(reward = 1.0, terminal = false, state = 1)

julia> new_obs = StateOverriddenObs(;obs=old_obs, state=nothing)
StateOverriddenObs{NamedTuple{(:reward, :terminal, :state),Tuple{Float64,Bool,Int64}},Nothing}((reward = 1.0, terminal = false, state = 1), nothing)

julia> get_state(new_obs) === nothing
true

julia> get_reward(new_obs) === get_reward(old_obs)
true

Trajectory

ReinforcementLearningBase.AbstractTrajectory — Type

AbstractTrajectory{names,types} <: AbstractArray{NamedTuple{names,types},1}

A trajectory is used to record some useful information during the interactions between agents and environments.

Parameters

names::NTuple{Symbol}, indicate what fields to be recorded.

types::Tuple{DataType...}, the datatypes of names.

The length of names and types must match.

Required Methods:

get_trace

Base.push!(t::AbstractTrajectory, kv::Pair{Symbol})

Base.pop!(t::AbstractTrajectory, s::Symbol)

Optional Methods:

Base.length

Base.size

Base.lastindex

Base.isempty

Base.empty!

Base.push!

Base.pop!

ReinforcementLearningBase.get_trace — Function

get_trace(t::AbstractTrajectory, s::Symbol)

get_trace(t::AbstractTrajectory, s::NTuple{N,Symbol}) where {N}

get_trace(t::AbstractTrajectory, s::Symbol...)

get_trace(t::AbstractTrajectory{names}) where {names}

Base.push! — Function

Base.push!(t::AbstractTrajectory; kwargs...)

Base.push!(t::AbstractTrajectory, kv::Pair{Symbol})

Base.pop! — Function

Base.pop!(t::AbstractTrajectory{names}) where {names}

pop! out one element of each trace in t

Base.pop!(t::AbstractTrajectory, s::Symbol...)

pop! out one element of the traces specified in s

Base.pop!(t::AbstractTrajectory, s::Symbol)

pop! out one element of the trace s in t

Space

ReinforcementLearningBase.AbstractSpace — Type

Describe the span of observations and actions. Usually the following methods are implemented:

Base.length

Base.in

Random.rand

Base.eltype

Some typical implementations of AbstractSpace are also included in the RLBase.

ReinforcementLearningBase.EmptySpace — Type

EmptySpace()

There's nothing in the EmptySpace!

ReinforcementLearningBase.DiscreteSpace — Type

DiscreteSpace(span)

The span can be of any iterators.

Example

julia> s = DiscreteSpace([1,2,3])
DiscreteSpace{Array{Int64,1}}([1, 2, 3])

julia> 0 ∉ s
true

julia> 2 ∈ s
true

julia> s = DiscreteSpace(Set([:a, :c, :a, :b]))
DiscreteSpace{Set{Symbol}}(Set(Symbol[:a, :b, :c]))

julia> s = DiscreteSpace(3)
DiscreteSpace{UnitRange{Int64}}(1:3)

ReinforcementLearningBase.MultiDiscreteSpace — Type

MultiDiscreteSpace(low::T, high::T) where {T<:AbstractArray}

Similar to DiscreteSpace, but scaled to multi-dimension.

ReinforcementLearningBase.ContinuousSpace — Type

ContinuousSpace(low, high)

Similar to DiscreteSpace, but the span is continuous.

ReinforcementLearningBase.MultiContinuousSpace — Type

MultiContinuousSpace(low, high)

Similar to ContinuousSpace, but scaled to multi-dimension.

ReinforcementLearningBase.TupleSpace — Type

TupleSpace(args::AbstractSpace...)

ReinforcementLearningBase.DictSpace — Type

DictSpace(ps::Pair{<:Union{Symbol,AbstractString},<:AbstractSpace}...)