ReinforcementLearningBase.jl

ReinforcementLearningBase.RLBase — Module

ReinforcementLearningBase.jl (RLBase) provides some common constants, traits, abstractions and interfaces in developing reinforcement learning algorithms in Julia. From the concept level, they can be organized in the following parts:

Policy

EnvironmentModel

Environment

Traits for Environment

Observation of Environment

Observation Space and Action Space

Policy

ReinforcementLearningBase.AbstractPolicy — Type

(π::AbstractPolicy)(obs) -> action

Policy is the most basic concept in reinforcement learning. A policy is a functional object which takes in an observation and generate an action.

ReinforcementLearningBase.update! — Method

update!(π::AbstractPolicy, experience)

Update the policy π with online/offline experience.

ReinforcementLearningBase.get_prob — Method

get_prob(π::AbstractPolicy, obs)

Get the probability distribution of actions based on policy π given an observation obs.

ReinforcementLearningBase.get_prob — Method

get_prob(π::AbstractPolicy, obs, action)

Only valid for environments with discrete action space.

ReinforcementLearningBase.get_priority — Method

get_priority(π::AbstractPolicy, experience)

Usually used in offline policies.

EnvironmentModel

ReinforcementLearningBase.AbstractEnvironmentModel — Type

Describe how to model a reinforcement learning environment. TODO: need more investigation Ref: https://bair.berkeley.edu/blog/2019/12/12/mbpo/

Analytic gradient computation

Sampling-based planning

Model-based data generation

Value-equivalence prediction

Environment

ReinforcementLearningBase.AbstractEnv — Type

(env::AbstractEnv)(action) = env(get_current_player(env), action) -> nothing
(env::AbstractEnv)(player, action) -> nothing

Super type of all reinforcement learning environments.

ReinforcementLearningBase.get_action_space — Function

get_action_space(env::AbstractEnv) -> AbstractSpace

Missing docstring.

Missing docstring for get_state_space. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_current_player. Check Documenter's build log for details.

ReinforcementLearningBase.get_player_id — Function

get_player_id(player, env::AbstractEnv) -> Int

Get the index of current player. Result should be an Int and starts from 1. Usually used in multi-agent environments.

ReinforcementLearningBase.get_num_players — Function

get_num_players(env::AbstractEnv) -> Int

ReinforcementLearningBase.reset! — Method

Reset the internal state of an environment

ReinforcementLearningBase.render — Function

Show the environment in a user-friendly manner

Missing docstring.

Missing docstring for Random.seed!(env::AbstractEnv, seed). Check Documenter's build log for details.

Base.copy — Method

Make an independent copy of env

ReinforcementLearningBase.get_history — Method

Get all actions in each ply

Traits of Environment

ReinforcementLearningBase.DynamicStyle — Function

DynamicStyle(env::AbstractEnv) = SEQUENTIAL

Determine whether the players can play simultaneous or not. Default value is SEQUENTIAL

ReinforcementLearningBase.ChanceStyle — Function

Either DETERMINISTIC or STOCHASTIC.

ReinforcementLearningBase.InformationStyle — Function

Specify if the env is PERFECT_INFORMATION or IMPERFECT_INFORMATION

ReinforcementLearningBase.RewardStyle — Function

Specify whether we can get reward after each step or only at the end of an game. Possible values are STEP_REWARD or TERMINAL_REWARD

ReinforcementLearningBase.UtilityStyle — Function

UtilityStyle(env::AbstractEnv)

Specify the utility style in multi-agent environments. Possible values are:

ZERO_SUM

CONSTANT_SUM

GENERAL_SUM

IDENTICAL_REWARD

ReinforcementLearningBase.ActionStyle — Function

ActionStyle(env::AbstractEnv)
ActionStyle(obs)

Specify whether the observation contains a full action set or a minimal action set. By default the MINIMAL_ACTION_SET is returned.

ReinforcementLearningBase.SEQUENTIAL — Constant

Environment with the DynamicStyle of SEQUENTIAL must takes actions from different players one-by-one.

ReinforcementLearningBase.SIMULTANEOUS — Constant

Environment with the DynamicStyle of SIMULTANEOUS must take in actions from some (or all) players at one time

ReinforcementLearningBase.DETERMINISTIC — Constant

Observations are solely determined by action

ReinforcementLearningBase.STOCHASTIC — Constant

Observations are stochastic given by the same action

ReinforcementLearningBase.PERFECT_INFORMATION — Constant

All players observe the same state

ReinforcementLearningBase.IMPERFECT_INFORMATION — Constant

The inner state of some players' observations may be different

ReinforcementLearningBase.ZERO_SUM — Constant

Rewards of all players sum to 0

ReinforcementLearningBase.CONSTANT_SUM — Constant

Rewards of all players sum to a constant

ReinforcementLearningBase.GENERAL_SUM — Constant

Total rewards of all players may be different in each step

ReinforcementLearningBase.IDENTICAL_REWARD — Constant

Every player gets the same reward

ReinforcementLearningBase.FULL_ACTION_SET — Constant

The action space of the environment may contains illegal actions

ReinforcementLearningBase.MINIMAL_ACTION_SET — Constant

All actions in the action space of the environment are legal

Several meta-environments are also provided:

ReinforcementLearningBase.WrappedEnv — Type

WrappedEnv(;preprocessor=identity, env, postprocessor=identity)

The observation of the inner env is first transformed by the preprocessor. And the action is transformed by postprocessor and then send to the inner env.

Observation

ReinforcementLearningBase.observe — Function

observe(env::AbstractEnv) = observe(env, get_current_player(env))
observe(::AbstractEnv, player)

Get an observation of the env from the perspective of an player.

Note

This is a very deliberate decision to adopt the duck-typing here to describe an observation from an environment. By default, we assume an observation is a NamedTuple, which is the most common case. But of course it can be of any type, as long as it implemented the necessay methods described in this section.

ReinforcementLearningBase.get_legal_actions_mask — Function

get_legal_actions_mask(obs) -> Bool[]

Only valid for observations of FULL_ACTION_SET.

ReinforcementLearningBase.get_legal_actions — Function

get_legal_actions(obs)

Only valid for observations of FULL_ACTION_SET.

ReinforcementLearningBase.get_terminal — Function

get_terminal(obs)

ReinforcementLearningBase.get_reward — Function

get_reward(obs)

ReinforcementLearningBase.get_state — Function

get_state(obs)

Space

ReinforcementLearningBase.AbstractSpace — Type

Describe the span of observations and actions. Usually the following methods are implemented:

Base.length

Base.in

Random.rand

Base.eltype

ReinforcementLearningBase.EmptySpace — Type

EmptySpace()

There's nothing in the EmptySpace!

ReinforcementLearningBase.DiscreteSpace — Type

DiscreteSpace(span)

The span can be of any iterators.

Example

julia> s = DiscreteSpace([1,2,3])
DiscreteSpace{Array{Int64,1}}([1, 2, 3])

julia> 0 ∉ s
true

julia> 2 ∈ s
true

julia> s = DiscreteSpace(Set([:a, :c, :a, :b]))
DiscreteSpace{Set{Symbol}}(Set(Symbol[:a, :b, :c]))

julia> s = DiscreteSpace(3)
DiscreteSpace{UnitRange{Int64}}(1:3)

ReinforcementLearningBase.MultiDiscreteSpace — Type

MultiDiscreteSpace(low::T, high::T) where {T<:AbstractArray}

Similar to DiscreteSpace, but scaled to multi-dimension.

ReinforcementLearningBase.ContinuousSpace — Type

ContinuousSpace(low, high)

Similar to DiscreteSpace, but the span is continuous.

ReinforcementLearningBase.MultiContinuousSpace — Type

MultiContinuousSpace(low, high)

Similar to ContinuousSpace, but scaled to multi-dimension.

ReinforcementLearningBase.TupleSpace — Type

TupleSpace(args::AbstractSpace...)

ReinforcementLearningBase.DictSpace — Type

DictSpace(ps::Pair{<:Union{Symbol,AbstractString},<:AbstractSpace}...)

Utils

Base.run — Method

run(π, env::AbstractEnv)

Run the policy π in env until the end.

ReinforcementLearningBase.RandomPolicy — Type

RandomPolicy(action_space, rng)

Construct a random policy with actions in action_space.

ReinforcementLearningBase.StateOverriddenObs — Type

StateOverriddenObs(;obs, state)

Replace the internal state of obs with state.

Example

julia> old_obs = (reward=1.0, terminal=false, state=1)
(reward = 1.0, terminal = false, state = 1)

julia> new_obs = StateOverriddenObs(;obs=old_obs, state=nothing)
StateOverriddenObs{NamedTuple{(:reward, :terminal, :state),Tuple{Float64,Bool,Int64}},Nothing}((reward = 1.0, terminal = false, state = 1), nothing)

julia> get_state(new_obs) === nothing
true

julia> get_reward(new_obs) === get_reward(old_obs)
true

ReinforcementLearningBase.BatchObs — Type

BatchObs(obs::Vector)

Wrap several observations into a batch.

Missing docstring.

Missing docstring for WrappedEnv. Check Documenter's build log for details.

ReinforcementLearningBase.MultiThreadEnv — Type

MultiThreadEnv(envs::Vector{<:AbstractEnv})

Wrap multiple environments in one environment. Each environment will run in parallel by leveraging Threads.@spawn.

ReinforcementLearningBase.AbstractPreprocessor — Type

(p::AbstractPreprocessor)(obs)

By default a StateOverriddenObs is returned to avoid modifying original observation.

ReinforcementLearningBase.CloneStatePreprocessor — Type

CloneStatePreprocessor()

Do deepcopy for the state in an observation.

ReinforcementLearningBase.ComposedPreprocessor — Type

ComposedPreprocessor(p::AbstractPreprocessor...)

Compose multiple preprocessors.