ReinforcementLearningBase.jl

ReinforcementLearningBase.RLBase — Module

ReinforcementLearningBase.jl (RLBase) provides some common constants, traits, abstractions and interfaces in developing reinforcement learning algorithms in Julia. From the concept level, they can be organized in the following parts:

Policy

EnvironmentModel

Environment

Traits for Environment

ReinforcementLearningBase.CONSTANT_SUM — Constant

Rewards of all players sum to a constant

ReinforcementLearningBase.DETERMINISTIC — Constant

No chance player in the environment. And the game is deterministic.

ReinforcementLearningBase.EXPLICIT_STOCHASTIC — Constant

Environment contains chance player and the probability is known.

ReinforcementLearningBase.FULL_ACTION_SET — Constant

The action space of the environment may contains illegal actions

ReinforcementLearningBase.GENERAL_SUM — Constant

Total rewards of all players may be different in each step

ReinforcementLearningBase.IDENTICAL_REWARD — Constant

Every player gets the same reward

ReinforcementLearningBase.IMPERFECT_INFORMATION — Constant

The inner state of some players' observations may be different

ReinforcementLearningBase.MINIMAL_ACTION_SET — Constant

All actions in the action space of the environment are legal

ReinforcementLearningBase.PERFECT_INFORMATION — Constant

All players observe the same state

ReinforcementLearningBase.SAMPLED_STOCHASTIC — Constant

Environment contains chance player and the probability is unknown. Usually only a dummy action is allowed in this case.

ReinforcementLearningBase.SEQUENTIAL — Constant

Environment with the DynamicStyle of SEQUENTIAL must takes actions from different players one-by-one.

ReinforcementLearningBase.SIMULTANEOUS — Constant

Environment with the DynamicStyle of SIMULTANEOUS must take in actions from some (or all) players at one time

ReinforcementLearningBase.STEP_REWARD — Constant

We can get reward after each step

ReinforcementLearningBase.STOCHASTIC — Constant

No chance player in the environment. And the game is stochastic.

ReinforcementLearningBase.TERMINAL_REWARD — Constant

Only get reward at the end of environment

ReinforcementLearningBase.ZERO_SUM — Constant

Rewards of all players sum to 0

ReinforcementLearningBase.AbstractEnv — Type

(env::AbstractEnv)(action) = env(action, get_current_player(env)) -> env
(env::AbstractEnv)(action, player) -> env

Super type of all reinforcement learning environments.

ReinforcementLearningBase.AbstractEnvironmentModel — Type

Describe how to model a reinforcement learning environment. TODO: need more investigation Ref: https://bair.berkeley.edu/blog/2019/12/12/mbpo/

Analytic gradient computation

Sampling-based planning

Model-based data generation

Value-equivalence prediction

Model-based Reinforcement Learning: A Survey. Tutorial on Model-Based Methods in Reinforcement Learning

ReinforcementLearningBase.AbstractPolicy — Type

(π::AbstractPolicy)(env) -> action

Policy is the most basic concept in reinforcement learning. A policy is a functional object which takes in an environemnt and generate an action.

ReinforcementLearningBase.AbstractSpace — Type

Describe the span of states and actions. Usually the following methods are implemented:

Base.length

Base.in

Random.rand

Base.eltype

ReinforcementLearningBase.ContinuousSpace — Method

ContinuousSpace(low, high)

Similar to DiscreteSpace, but the span is continuous.

ReinforcementLearningBase.DictSpace — Method

DictSpace(ps::Pair{<:Union{Symbol,AbstractString},<:AbstractSpace}...)

ReinforcementLearningBase.DiscreteSpace — Type

DiscreteSpace(span)

The span can be of any iterators.

Example

julia> s = DiscreteSpace([1,2,3])
DiscreteSpace{Array{Int64,1}}([1, 2, 3])

julia> 0 ∉ s
true

julia> 2 ∈ s
true

julia> s = DiscreteSpace(Set([:a, :c, :a, :b]))
DiscreteSpace{Set{Symbol}}(Set(Symbol[:a, :b, :c]))

julia> s = DiscreteSpace(3)
DiscreteSpace{UnitRange{Int64}}(1:3)

ReinforcementLearningBase.DiscreteSpace — Method

DiscreteSpace(low, high)

Create a DiscreteSpace with span of low:high

ReinforcementLearningBase.DiscreteSpace — Method

DiscreteSpace(high::T)

Create a DiscreteSpace with span of 1:high

ReinforcementLearningBase.EmptySpace — Type

EmptySpace()

There's nothing in the EmptySpace!

ReinforcementLearningBase.MultiContinuousSpace — Method

MultiContinuousSpace(low, high)

Similar to ContinuousSpace, but scaled to multi-dimension.

ReinforcementLearningBase.MultiDiscreteSpace — Type

MultiDiscreteSpace(low::T, high::T) where {T<:AbstractArray}

Similar to DiscreteSpace, but scaled to multi-dimension.

ReinforcementLearningBase.MultiDiscreteSpace — Method

MultiDiscreteSpace(high::T) where {T<:AbstractArray}

The low will fall back to ones(eltype(T), size(high)).

ReinforcementLearningBase.MultiThreadEnv — Type

MultiThreadEnv(envs::Vector{<:AbstractEnv})

Wrap multiple environments into one environment. Each environment will run in parallel by leveraging Threads.@spawn. So remember to set the environment variable JULIA_NUM_THREADS!

ReinforcementLearningBase.RandomPolicy — Type

RandomPolicy(action_space, rng=Random.GLOBAL_RNG)

Construct a random policy with actions in action_space. If action_space is nothing then the legal_actions at runtime will be used to randomly sample a valid action.

ReinforcementLearningBase.RandomPolicy — Method

RandomPolicy(env::AbstractEnv; rng=Random.GLOBAL_RNG)

If env is of FULL_ACTION_SET, then action is randomly chosen at runtime in get_actions(env). Otherwise, the env is supposed to be of MINIMAL_ACTION_SET. The get_actions(env) is supposed to be static and will only be used to initialize the random policy for once.

ReinforcementLearningBase.StateCachedEnv — Type

Cache the state so that get_state(env) will always return the same result before the next interaction with env.

Base.copy — Method

Make an independent copy of env

Base.run — Method

run(π, env::AbstractEnv)

Run the policy π in env until the end.

Random.seed! — Method

Set the seed of internal rng

ReinforcementLearningBase.ActionStyle — Method

ActionStyle(env::AbstractEnv)

Specify whether the current state of env contains a full action set or a minimal action set. By default the MINIMAL_ACTION_SET is returned.

ReinforcementLearningBase.ChanceStyle — Method

ChanceStyle(env) = DETERMINISTIC

ReinforcementLearningBase.DynamicStyle — Method

DynamicStyle(env::AbstractEnv) = SEQUENTIAL

Determine whether the players can play simultaneously or not. Default value is SEQUENTIAL

ReinforcementLearningBase.InformationStyle — Method

InformationStyle(env) = PERFECT_INFORMATION

Specify whether the env is PERFECT_INFORMATION or IMPERFECT_INFORMATION. Return PERFECT_INFORMATION by default.

ReinforcementLearningBase.NumAgentStyle — Method

NumAgentStyle(env)

ReinforcementLearningBase.RewardStyle — Method

Specify whether we can get reward after each step or only at the end of an game. Possible values are STEP_REWARD or TERMINAL_REWARD

ReinforcementLearningBase.UtilityStyle — Method

UtilityStyle(env::AbstractEnv)

Specify the utility style in multi-agent environments. Possible values are:

ZERO_SUM

CONSTANT_SUM

GENERAL_SUM

IDENTICAL_REWARD

ReinforcementLearningBase.child — Method

child(env::AbstractEnv, action)

Treat the env as a game tree. Create an independent child after applying action.

ReinforcementLearningBase.get_actions — Function

get_actions(env, player=get_current_player(env))

Get all available actions from environment. See also: get_legal_actions

ReinforcementLearningBase.get_chance_player — Method

get_chance_player(env)

Only valid for environments with a chance player.

ReinforcementLearningBase.get_current_player — Method

get_current_player(env)

Return the next player to take action. For Extensive Form Games, A chance player may be returned. (See also get_chance_player) For SIMULTANEOUS environments, a simultaneous player may be returned. (See also get_simultaneouse_player).

ReinforcementLearningBase.get_history — Function

Get all actions in each ply

ReinforcementLearningBase.get_legal_actions — Function

get_legal_actions(env, player=get_current_player(env))

Only valid for environments of FULL_ACTION_SET.

ReinforcementLearningBase.get_legal_actions_mask — Function

get_legal_actions_mask(env, player=get_current_player(env)) -> AbstractArray{Bool}

Required for environments of FULL_ACTION_SET.

ReinforcementLearningBase.get_priority — Method

get_priority(π::AbstractPolicy, experience)

Usually used in offline policies.

ReinforcementLearningBase.get_prob — Function

get_prob(env, player=get_chance_player(env))

Only valid for environments of EXPLICIT_STOCHASTIC style. Here player must be a chance player.

ReinforcementLearningBase.get_prob — Method

get_prob(π::AbstractPolicy, env, action)

Only valid for environments with discrete action space.

ReinforcementLearningBase.get_prob — Method

get_prob(π::AbstractPolicy, env)

Get the probability distribution of actions based on policy π given an env.

ReinforcementLearningBase.get_reward — Function

get_reward(env, player=get_current_player(env))

ReinforcementLearningBase.get_simultaneouse_player — Method

get_simultaneouse_player(env)

Only valid for environments of SIMULTANEOUS style.

ReinforcementLearningBase.get_spectator_player — Method

get_spectator_player(env)

Used in imperfect multi-agent environments.

ReinforcementLearningBase.get_state — Function

get_state([t::Type], env, player=get_current_player(env)) -> state

The state can be of any type. However, most neural network based algorithms assume it's an AbstractArray. For environments with many different states provided (inner state, information state, etc), users need to provide t::Type to declare which kind of state they want.

ReinforcementLearningBase.get_terminal — Function

get_terminal(env, player=get_current_player(env))

ReinforcementLearningBase.reset! — Method

Reset the internal state of an environment

ReinforcementLearningBase.update! — Method

update!(π::AbstractPolicy, experience)

Update the policy π with online/offline experience.