ReinforcementLearningEnvironments.jl

ReinforcementLearningEnvironments.jl (RLEnvs) provides some very common environments (for example: CartPoleEnv) together with some wrappers for 3-rd party environments.

To use those extra environments with ReinforcementLearning.jl, you have to manually install those dependent packages listed here. And use it like this:

julia> using ArcadeLearningEnvironment
┌ Warning: Package ReinforcementLearningEnvironments does not have ArcadeLearningEnvironment in its dependencies:
│ - If you have ReinforcementLearningEnvironments checked out for development and have
│   added ArcadeLearningEnvironment as a dependency but haven't updated your primary
│   environment's manifest file, try `Pkg.resolve()`.
│ - Otherwise you may need to report an issue with ReinforcementLearningEnvironments
└ Loading ArcadeLearningEnvironment into ReinforcementLearningEnvironments from project dependency, future warnings for ReinforcementLearningEnvironments are suppressed.

julia> using ReinforcementLearning.ReinforcementLearningEnvironments

julia> env = AtariEnv()

The warning above can be safely ignored. Especially notice that you need to execute using ReinforcementLearning.ReinforcementLearningEnvironments.

ReinforcementLearningEnvironments.CartPoleEnv — Type

CartPoleEnv(;kwargs...)

Keyword arguments

T = Float64

gravity = T(9.8)

masscart = T(1.0)

masspole = T(0.1)

halflength = T(0.5)

forcemag = T(10.0)

max_steps = 200

seed = nothing

ReinforcementLearningEnvironments.MountainCarEnv — Type

MountainCarEnv(;kwargs...)

Keyword arguments

T = Float64

continuous = false

seed = nothing

min_pos = -1.2

max_pos = 0.6

max_speed = 0.07

goal_pos = 0.5

max_steps = 200

goal_velocity = 0.0

power = 0.001

gravity = 0.0025

ReinforcementLearningEnvironments.PendulumEnv — Type

PwendulumEnv(;kwargs...)

Keyword arguments

T = Float64

max_speed = T(8)

max_torque = T(2)

g = T(10)

m = T(1)

l = T(1)

dt = T(0.05)

max_steps = 200

seed = nothing

ReinforcementLearningEnvironments.PendulumNonInteractiveEnv — Type

A non-interactive pendulum environment.

Accepts only nothing actions, which result in the system being simulated for one time step. Sets env.done to true once maximum_time is reached. Resets to a random position and momentum. Always returns zero rewards.

Useful for debugging and development purposes, particularly in model-based reinforcement learning.