ReinforcementLearningZoo.jl

ReinforcementLearningZoo.A2CGAELearner — Type

A2CGAELearner(;kwargs...)

Keyword arguments

approximator, an ActorCritic based NeuralNetworkApproximator

γ::Float32, reward discount rate.

'λ::Float32', lambda for GAE-lambda

actor_loss_weight::Float32

critic_loss_weight::Float32

entropy_loss_weight::Float32

ReinforcementLearningZoo.A2CLearner — Type

A2CLearner(;kwargs...)

Keyword arguments

approximator::ActorCritic

γ::Float32, reward discount rate.

actor_loss_weight::Float32

critic_loss_weight::Float32

entropy_loss_weight::Float32

ReinforcementLearningZoo.BasicDQNLearner — Type

BasicDQNLearner(;kwargs...)

See paper: Playing Atari with Deep Reinforcement Learning

This is the very basic implementation of DQN. Compared to the traditional Q learning, the only difference is that, in the updating step it uses a batch of transitions sampled from an experience buffer instead of current transition. And the approximator is usually a NeuralNetworkApproximator. You can start from this implementation to understand how everything is organized and how to write your own customized algorithm.

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

loss_func: the loss function to use. TODO: provide a default huber_loss?

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

rng=Random.GLOBAL_RNG

ReinforcementLearningZoo.DDPGPolicy — Method

DDPGPolicy(;kwargs...)

Keyword arguments

behavior_actor,

behavior_critic,

target_actor,

target_critic,

start_policy,

γ = 0.99f0,

ρ = 0.995f0,

batch_size = 32,

start_steps = 10000,

update_after = 1000,

update_every = 50,

act_limit = 1.0,

act_noise = 0.1,

step = 0,

rng = Random.GLOBAL_RNG,

ReinforcementLearningZoo.DQNLearner — Type

DQNLearner(;kwargs...)

See paper: Human-level control through deep reinforcement learning

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

rng = Random.GLOBAL_RNG

ReinforcementLearningZoo.DQNLearner — Method

Note

The state of the observation is assumed to have been stacked, if !isnothing(stack_size).

ReinforcementLearningZoo.IQNLearner — Type

IQNLearner(;kwargs)

See paper

Keyworkd arugments

approximator, a ImplicitQuantileNet

target_approximator, a ImplicitQuantileNet, must have the same structure as approximator

κ = 1.0f0,

N = 32,

N′ = 32,

Nₑₘ = 64,

K = 32,

γ = 0.99f0,

stack_size = 4,

batch_size = 32,

update_horizon = 1,

min_replay_history = 20000,

update_freq = 4,

target_update_freq = 8000,

update_step = 0,

default_priority = 1.0f2,

β_priority = 0.5f0,

rng = Random.GLOBAL_RNG,

device_seed = nothing,

ReinforcementLearningZoo.ImplicitQuantileNet — Type

ImplicitQuantileNet(;ψ, ϕ, header)

        quantiles (n_action, n_quantiles, batch_size)
           ↑
         header
           ↑
feature ↱  ⨀   ↰ transformed embedding
       ψ       ϕ
       ↑       ↑
       s        τ

ReinforcementLearningZoo.PPOLearner — Type

PPOLearner(;kwargs)

Keyword arguments

approximator,

γ = 0.99f0,

λ = 0.95f0,

clip_range = 0.2f0,

max_grad_norm = 0.5f0,

n_microbatches = 4,

n_epochs = 4,

actor_loss_weight = 1.0f0,

critic_loss_weight = 0.5f0,

entropy_loss_weight = 0.01f0,

rng = Random.GLOBAL_RNG,

ReinforcementLearningZoo.PrioritizedDQNLearner — Type

PrioritizedDQNLearner(;kwargs...)

See paper: Prioritized Experience Replay And also https://danieltakeshi.github.io/2019/07/14/per/

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

rng = Random.GLOBAL_RNG

ReinforcementLearningZoo.PrioritizedDQNLearner — Method

Note

The state of the observation is assumed to have been stacked, if !isnothing(stack_size).

ReinforcementLearningZoo.RainbowLearner — Type

RainbowLearner(;kwargs...)

See paper: Rainbow: Combining Improvements in Deep Reinforcement Learning

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

Vₘₐₓ::Float32: the maximum value of distribution.

Vₘᵢₙ::Float32: the minimum value of distribution.

n_actions::Int: number of possible actions.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=500: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float32=1.0f2.: the default priority for newly added transitions. It must be >= 1.

n_atoms::Int=51: the number of buckets of the value function distribution.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

rng = Random.GLOBAL_RNG