ReinforcementLearningZoo.jl

ReinforcementLearningZoo.jl (RLZoo) provides some implementations of the most typical reinforcement learning algorithms.

ReinforcementLearningZoo.BasicDQNLearner — Type

BasicDQNLearner(;kwargs...)

See paper: Playing Atari with Deep Reinforcement Learning

This is the very basic implementation of DQN. Compared to the traditional Q learning, the only difference is that, in the updating step it uses a batch of transitions sampled from an experience buffer instead of current transition. And the approximator is usually a NeuralNetworkApproximator. You can start from this implementation to understand how everything is organized and how to write your own customized algorithm.

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

loss_func: the loss function to use. TODO: provide a default huber_loss?

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

seed=nothing.

ReinforcementLearningZoo.DQNLearner — Type

DQNLearner(;kwargs...)

See paper: Human-level control through deep reinforcement learning

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

seed = nothing

ReinforcementLearningZoo.PrioritizedDQNLearner — Type

PrioritizedDQNLearner(;kwargs...)

See paper: Prioritized Experience Replay

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=100: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

seed = nothing

ReinforcementLearningZoo.RainbowLearner — Type

RainbowLearner(;kwargs...)

See paper: Rainbow: Combining Improvements in Deep Reinforcement Learning

Keywords

approximator::AbstractApproximator: used to get Q-values of a state.

target_approximator::AbstractApproximator: similar to approximator, but used to estimate the target (the next state).

loss_func: the loss function.

Vₘₐₓ::Float32: the maximum value of distribution.

Vₘᵢₙ::Float32: the minimum value of distribution.

n_actions::Int: number of possible actions.

γ::Float32=0.99f0: discount rate.

batch_size::Int=32

update_horizon::Int=1: length of update ('n' in n-step update).

min_replay_history::Int=32: number of transitions that should be experienced before updating the approximator.

update_freq::Int=4: the frequency of updating the approximator.

target_update_freq::Int=500: the frequency of syncing target_approximator.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

n_atoms::Int=51: the number of buckets of the value function distribution.

stack_size::Union{Int, Nothing}=4: use the recent stack_size frames to form a stacked state.

default_priority::Float64=100.: the default priority for newly added transitions.

seed = nothing